Trist warnings (#126)
Addressed the compiler warnings in fakeVis and timeDownsample, as requested in #88 .

* fixed warnings in timeDownsample

* fixed warnings in fakeVis

* changed type of iterations to uint32_t

+++ lib/processes/timeDownsample.cpp
@@ -66,7 +66,7 @@ void timeDownsample::main_thread() {
             // Set the parameters of the accumulation window
             wdw_len = nsamp * frame.fpga_seq_length;
             wdw_end = fpga_seq_start + wdw_len;
-        } else if (frame.freq_id != freq_id) {
+        } else if (frame.freq_id != (unsigned) freq_id) {
             throw std::runtime_error("Cannot downsample stream with more than one frequency.");
         }
 
@@ -120,9 +120,9 @@ void timeDownsample::main_thread() {
                 // average inverse weights, i.e. variance
                 output_frame.weight[i] += 1. / frame.weight[i];
             }
-            for (int i = 0; i < num_eigenvectors; i++) {
+            for (uint32_t i = 0; i < num_eigenvectors; i++) {
                 output_frame.eval[i] += frame.eval[i];
-                for (int j = 0; j < num_elements; j++) {
+                for (uint32_t j = 0; j < num_elements; j++) {
                     int k = i * num_elements + j;
                     output_frame.evec[k] += frame.evec[k];
                 }
@@ -144,9 +144,9 @@ void timeDownsample::main_thread() {
                 // extra factor of nsamp for sample variance
                 output_frame.weight[i] = nframes*nframes / output_frame.weight[i];
             }
-            for (int i = 0; i < num_eigenvectors; i++) {
+            for (uint32_t i = 0; i < num_eigenvectors; i++) {
                 output_frame.eval[i] /= nframes;
-                for (int j = 0; j < num_elements; j++) {
+                for (uint32_t j = 0; j < num_elements; j++) {
                     int k = i * num_elements + j;
                     output_frame.evec[k] /= nframes;
                 }
